Remove injectDefinitions import, prefer injector
