+++ packages/api-metadata/src/consts/fromMetadata/fromMetadata.spec.ts
@@ -2,13 +2,12 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
+import '@polkadot/types/injector';
+
 import { ClassOf, Metadata } from '@polkadot/types';
 import json from '@polkadot/types/Metadata/static';
-import { injectDefinitions } from '@polkadot/types/srml';
 
 import fromMetadata from '../fromMetadata';
-
-injectDefinitions();
 
 // Use the pre-generated metadata
 const metadata = new Metadata(json);
