+++ packages/api-metadata/src/extrinsics/fromMetadata/fromMetadata.spec.ts
@@ -2,13 +2,12 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
+import '@polkadot/types/injector';
+
 import { Extrinsic, Metadata } from '@polkadot/types';
-import { injectDefinitions } from '@polkadot/types/srml';
 import json from '@polkadot/types/Metadata/v7/static';
 
 import fromMetadata from '.';
-
-injectDefinitions();
 
 // Use the pre-generated metadata
 const metadata = new Metadata(json);
