+++ packages/api-metadata/src/extrinsics/index.spec.ts
@@ -2,8 +2,9 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
+import '@polkadot/types/injector';
+
 import testingPairs from '@polkadot/keyring/testingPairs';
-import { injectDefinitions } from '@polkadot/types/srml';
 import { Extrinsic } from '@polkadot/types';
 
 import extrinsics from './static';
@@ -11,10 +12,6 @@ import extrinsics from './static';
 const keyring = testingPairs({ type: 'ed25519' }, false);
 
 describe('extrinsics', (): void => {
-  beforeEach((): void => {
-    injectDefinitions();
-  });
-
   it('encodes an actual transfer (actual data)', (): void => {
     expect(
       new Extrinsic(
