+++ packages/api-metadata/src/storage/fromMetadata/fromMetadata.spec.ts
@@ -2,15 +2,14 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
+import '@polkadot/types/injector';
+
 import testingPairs from '@polkadot/keyring/testingPairs';
 import { Metadata } from '@polkadot/types';
 import json from '@polkadot/types/Metadata/static';
-import { injectDefinitions } from '@polkadot/types/srml';
 import { u8aToHex } from '@polkadot/util';
 
 import fromMetadata from '../fromMetadata';
-
-injectDefinitions();
 
 const keyring = testingPairs({ type: 'ed25519' });
