+++ packages/types/src/codec/Struct.spec.ts
@@ -2,20 +2,17 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
+import '../injector';
+
 import { ClassOf } from '../codec/createType';
 import AccountId from '../primitive/AccountId';
 import Text from '../primitive/Text';
 import U32 from '../primitive/U32';
-import { injectDefinitions } from '../srml';
 import { CodecTo } from '../types';
 import Struct from './Struct';
 import Vec from './Vec';
 
 describe('Struct', (): void => {
-  beforeEach((): void => {
-    injectDefinitions();
-  });
-
   describe('decoding', (): void => {
     const testDecode = (type: string, input: any): void =>
       it(`can decode from ${type}`, (): void => {
