+++ packages/types/src/codec/Tuple.spec.ts
@@ -2,6 +2,8 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
+import '../injector';
+
 import { VoteThreshold } from '../srml/elections/types';
 import { BlockNumber } from '../srml/runtime/types';
 import { CodecTo } from '../types';
@@ -12,14 +14,12 @@ import { ClassOf } from '../codec/createType';
 import Method from '../primitive/Method';
 import Text from '../primitive/Text';
 import U32 from '../primitive/U32';
-import { injectDefinitions } from '../srml';
 import Tuple from './Tuple';
 
 describe('Tuple', (): void => {
   let tuple: Tuple;
 
   beforeEach((): void => {
-    injectDefinitions();
     tuple = new Tuple(
       [Text, U32],
       ['bazzing', 69]
