+++ packages/types/src/primitive/AccountId.spec.ts
@@ -2,20 +2,17 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
+import '../injector';
+
 import { setAddressPrefix } from '@polkadot/util-crypto';
 
 import createType from '../codec/createType';
 import U8a from '../codec/U8a';
 import Vec from '../codec/Vec';
 import jsonVec from '../json/AccountIdVec.001.json';
-import { injectDefinitions } from '../srml';
 import AccountId from './AccountId';
 
 describe('AccountId', (): void => {
-  beforeEach((): void => {
-    injectDefinitions();
-  });
-
   describe('defaults', (): void => {
     const id = new AccountId();
