+++ packages/types/src/primitive/Block.spec.ts
@@ -2,16 +2,16 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
+import '../injector';
+
 import extrinsics from '@polkadot/api-metadata/extrinsics/static';
 
 import block00300 from '../json/SignedBlock.003.00.json';
-import { injectDefinitions } from '../srml';
 import Method from './Method';
 import Block from './Block';
 
 describe('Block', (): void => {
   beforeEach((): void => {
-    injectDefinitions();
     Method.injectMethods(extrinsics);
   });
