+++ packages/types/src/primitive/Extrinsic/v2/ExtrinsicExtra.spec.ts
@@ -2,14 +2,11 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
-import { injectDefinitions } from '../../../srml';
+import '../../../injector';
+
 import ExtrinsicExtra from './ExtrinsicExtra';
 
 describe('ExtrinsicExtraV2', (): void => {
-  beforeEach((): void => {
-    injectDefinitions();
-  });
-
   it('encodes to a sane Uint8Array (default construction)', (): void => {
     expect(
       new ExtrinsicExtra().toU8a()
