+++ packages/types/src/primitive/Extrinsic/v2/ExtrinsicSignature.spec.ts
@@ -2,14 +2,11 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
-import { injectDefinitions } from '../../../srml';
+import '../../../injector';
+
 import ExtrinsicSignature from './ExtrinsicSignature';
 
 describe('ExtrinsicSignatureV2', (): void => {
-  beforeEach((): void => {
-    injectDefinitions();
-  });
-
   it('encodes to a sane Uint8Array', (): void => {
     const u8a = new Uint8Array([
       // signer as an AccountIndex
