+++ packages/types/src/srml/rpc/ChainProperties.spec.ts
@@ -2,14 +2,11 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
+import '../../injector';
+
 import createType from '../../codec/createType';
-import { injectDefinitions } from '..';
 
 describe('ChainProperties', (): void => {
-  beforeEach((): void => {
-    injectDefinitions();
-  });
-
   it('decodes from a null value', (): void => {
     expect(
       [...createType('ChainProperties', null).entries()]
