+++ packages/types/src/srml/rpc/ExtrinsicStatus.spec.ts
@@ -2,18 +2,17 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
+import '../../injector';
+
 import { ExtrinsicStatus } from './types';
 
 import createType from '../../codec/createType';
 import rpc from '../../json/ExtrinsicStatus.001.json';
-import { injectDefinitions } from '..';
 
 describe('ExtrinsicStatus', (): void => {
   let status: ExtrinsicStatus;
 
   beforeEach((): void => {
-    injectDefinitions();
-
     status = createType('ExtrinsicStatus', rpc.params.result);
   });
