+++ packages/types/src/srml/rpc/RuntimeVersion.spec.ts
@@ -2,17 +2,17 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
+import '../../injector';
+
 import { RuntimeVersion } from './types';
 
 import createType from '../../codec/createType';
 import rpc from '../../json/RuntimeVersion.002.json';
-import { injectDefinitions } from '..';
 
 describe('RuntimeVersion', (): void => {
   let version: RuntimeVersion;
 
   beforeEach((): void => {
-    injectDefinitions();
     version = createType('RuntimeVersion', rpc.result);
   });
