+++ packages/types/src/srml/runtime/KeyValue.spec.ts
@@ -2,16 +2,13 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
+import '../../injector';
+
 import { KeyValue } from './types';
 
 import createType from '../../codec/createType';
-import { injectDefinitions } from '..';
 
 describe('KeyValue', (): void => {
-  beforeEach((): void => {
-    injectDefinitions();
-  });
-
   it('decodes KeyValue from u8a', (): void => {
     expect(
       createType('KeyValue', Uint8Array.from([
