+++ packages/types/src/srml/system/EventRecord.spec.ts
@@ -2,23 +2,20 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
+import '../../injector';
+
 import { EventRecord } from './types';
 
 import createType from '../../codec/createType';
 import Vec from '../../codec/Vec';
 import json1 from '../../json/EventRecord.001.json';
 import json3 from '../../json/EventRecord.003.json';
-import { injectDefinitions } from '..';
 import Metadata from '../../Metadata';
 import metadataV0 from '../../Metadata/v0/static';
 import metadata from '../../Metadata/static';
 import Event from '../../primitive/Event';
 
 describe('EventRecord', (): void => {
-  beforeEach((): void => {
-    injectDefinitions();
-  });
-
   describe('EventRecord_0_76', (): void => {
     beforeEach((): void => {
       Event.injectMetadata(
@@ -36,8 +33,6 @@ describe('EventRecord', (): void => {
 
   describe('EventRecord (current)', (): void => {
     beforeEach((): void => {
-      injectDefinitions();
-
       Event.injectMetadata(
         new Metadata(metadata)
       );
